Standardised Abstract classes and renamed pages to views
