+++ frontend/static/js/views/dashboardView.js
