+++ frontend/static/js/views/friendsView.js
