+++ frontend/static/js/views/landingView.js
