+++ frontend/static/js/views/leaderboardView.js
