+++ frontend/static/js/views/loginView.js
