+++ frontend/static/js/views/logoutView.js
