+++ frontend/static/js/views/playView.js
