+++ frontend/static/js/views/registerView.js
